refactor: 마이페이지 프로필 사진 컴포넌트화
