+++ src/models/MyPage.ts
@@ -1,0 +1,4 @@
+export interface ProfileImageProps {
+  profileImage: string;
+  patchUserProfileImage: (file: File) => void;
+}
