+++ src/pages/MyPage/index.tsx
@@ -1,10 +1,5 @@
 import React, { useState } from 'react';
-import defaultImg from '../../assets/icons/logo_black.svg';
-import {
-  BsImageFill,
-  BsPencilFill,
-  BsFillCheckCircleFill,
-} from 'react-icons/bs';
+import { BsPencilFill, BsFillCheckCircleFill } from 'react-icons/bs';
 import { Input } from '../../components/common/Input';
 import {
   EditInfo,
@@ -12,7 +7,6 @@ import {
   MyPageContent,
   MyPageHeader,
   MyPageHeaderUserName,
-  ProfileImage,
   ProfileInfo,
   ProfileInfoDetails,
   ProfileInfoDetailsContent,
@@ -24,13 +18,12 @@ import {
 } from './style';
 import { useUserProfile } from '../../hooks/Auth/useUserProfile';
 import { useUpdateProfile } from '../../hooks/Auth/useUpdateProfile';
+import { ProfileImg } from '../../components/myPage/ProfileImage';
 
 const MyPage: React.FC = () => {
   const { data, isLoading, error, refetch } = useUserProfile();
   const { patchUserNickName, patchUserLoginId, patchUserProfileImage } =
     useUpdateProfile();
-
-  const profileImageUploadRef = React.useRef<HTMLInputElement>(null);
 
   const [isEditingId, setIsEditingId] = useState(false);
   const [id, setId] = useState(data?.loginId ?? '');
@@ -42,23 +35,6 @@ const MyPage: React.FC = () => {
 
   const handleThemeChange = (theme: string) => {
     setSelectedTheme(theme);
-  };
-
-  const handleProfileImageUploadButtonClick = () => {
-    profileImageUploadRef.current?.click();
-  };
-
-  const handleProfileImageChange = async (
-    event: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    if (event.target.files && event.target.files[0]) {
-      try {
-        const file = event.target.files[0];
-        await patchUserProfileImage(file);
-      } catch (error) {
-        console.error('프로필 이미지 수정 실패:', error);
-      }
-    }
   };
 
   const handleNickNameChange = async () => {
@@ -103,23 +79,10 @@ const MyPage: React.FC = () => {
             사용자 정보
           </h2>
 
-          <ProfileImage>
-            <img src={data?.profileImage || defaultImg} alt="프로필 이미지" />
-            <button
-              aria-label="프로필 이미지 수정"
-              onClick={handleProfileImageUploadButtonClick}
-            >
-              <input
-                type="file"
-                accept="image/*"
-                id="profileImageUpload"
-                onChange={handleProfileImageChange}
-                ref={profileImageUploadRef}
-                style={{ display: 'none' }}
-              />
-              <BsImageFill className="icon_image" />
-            </button>
-          </ProfileImage>
+          <ProfileImg
+            profileImage={data?.profileImage ?? ''}
+            patchUserProfileImage={patchUserProfileImage}
+          />
 
           <ProfileInfo>
             <ProfileInfoDetails>
