+++ src/pages/MyPage/style.ts
@@ -39,41 +39,6 @@ export const UserInfoSection = styled.section`
   gap: 2.56rem;
 `;
 
-export const ProfileImage = styled.div`
-  width: 21.875rem;
-  height: 21.875rem;
-  border-radius: 1.0625rem;
-  overflow: hidden;
-  position: relative;
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  button {
-    width: 2.5rem;
-    height: 2.5rem;
-    border-radius: 0.4375rem;
-    background: var(--gray);
-    border: none;
-    position: absolute;
-    top: 18.25rem;
-    right: 1.13rem;
-    z-index: 2;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-
-    .icon_image {
-      width: 1.5rem;
-      height: 1.5rem;
-    }
-  }
-`;
-
 export const ProfileInfo = styled.div`
   display: flex;
   flex-direction: column;
